backend/requirements.txt
Flask
Flask-Login
werkzeug
mysql-connector-python